Increase tolerance to accommodate small inaccuracies in sin(). (#276)
Unfortunately the package does not pass its tests with mingw-w64 v12 on Windows. Rtools45 still uses mingw-w64 v11 to allow time for packages to be fixed, so one needs to use an unreleased version to see this.

The problem is that v12 decided to switch many C math functions to UCRT from previously used internal implementation. In case of s2, the failure is caused by very small inaccuracies in sin() - they are within 1 ULP, so programs should be able to work with that (the C standard doesn't give any limits to accuracy of math functions).